# frontend/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
...
    "/api/v1/symbols/{id}" : {
      "get" : {
        "tags" : [ "symbol-controller" ],
        "operationId" : "get_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SymbolDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/symbols/{id}/usages" : {
      "get" : {
        "tags" : [ "symbol-controller" ],
        "operationId" : "usages",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "confidence",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "EXACT", "RECOVERED", "NAME_ONLY" ]
            },
            "uniqueItems" : true
          }
        }, {
          "name" : "kind",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "CALL", "INSTANTIATION", "METHOD_REF", "TYPE_REF", "EXTENDS", "IMPLEMENTS", "OVERRIDES", "FIELD_READ", "FIELD_WRITE", "ANNOTATION" ]
            },
            "uniqueItems" : true
          }
        }, {
          "name" : "repo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageUsageView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/symbols/{id}/usages/summary" : {
      "get" : {
        "tags" : [ "symbol-controller" ],
        "operationId" : "summary",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UsageSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/symbols/search" : {
      "get" : {
        "tags" : [ "symbol-controller" ],
        "operationId" : "search",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "kind",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "CLASS", "INTERFACE", "ENUM", "RECORD", "ANNOTATION_TYPE", "METHOD", "CONSTRUCTOR", "FIELD" ]
          }
        }, {
          "name" : "repo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageSymbolHit"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "RepositoryRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "projectKey" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          }
        }
      },
...
      "DeclarationView" : {
        "type" : "object",
        "properties" : {
          "repository" : {
            "$ref" : "#/components/schemas/RepositoryRef"
          },
          "modulePath" : {
            "type" : "string"
          },
          "filePath" : {
            "type" : "string"
          },
          "line" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "SymbolDetail" : {
        "type" : "object",
        "properties" : {
          "symbol" : {
            "$ref" : "#/components/schemas/SymbolRef"
          },
          "origin" : {
            "type" : "string",
            "enum" : [ "SOURCE", "BINARY" ]
          },
          "nameOnly" : {
            "type" : "boolean"
          },
          "parent" : {
            "$ref" : "#/components/schemas/SymbolRef"
          },
          "declarations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DeclarationView"
            }
          },
          "members" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SymbolRef"
            }
          },
          "overrides" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SymbolRef"
            }
          },
          "overriddenBy" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SymbolRef"
            }
          },
          "supertypes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SymbolRef"
            }
          },
          "subtypes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SymbolRef"
            }
          }
        }
      },
      "SymbolRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "key" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "CLASS", "INTERFACE", "ENUM", "RECORD", "ANNOTATION_TYPE", "METHOD", "CONSTRUCTOR", "FIELD" ]
          },
          "display" : {
            "type" : "string"
          }
        }
      },
      "PageUsageView" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UsageView"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "UsageView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "from" : {
            "$ref" : "#/components/schemas/SymbolRef"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "CALL", "INSTANTIATION", "METHOD_REF", "TYPE_REF", "EXTENDS", "IMPLEMENTS", "OVERRIDES", "FIELD_READ", "FIELD_WRITE", "ANNOTATION" ]
          },
          "confidence" : {
            "type" : "string",
            "enum" : [ "EXACT", "RECOVERED", "NAME_ONLY" ]
          },
          "repository" : {
            "$ref" : "#/components/schemas/RepositoryRef"
          },
          "modulePath" : {
            "type" : "string"
          },
          "filePath" : {
            "type" : "string"
          },
          "line" : {
            "type" : "integer",
            "format" : "int32"
          },
          "column" : {
            "type" : "integer",
            "format" : "int32"
          },
          "snippet" : {
            "type" : "string"
          }
        }
      },
      "ClassUsage" : {
        "type" : "object",
        "properties" : {
          "classFqn" : {
            "type" : "string"
          },
          "usages" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ModuleUsage" : {
        "type" : "object",
        "properties" : {
          "modulePath" : {
            "type" : "string"
          },
          "usages" : {
            "type" : "integer",
            "format" : "int64"
          },
          "classes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ClassUsage"
            }
          }
        }
      },
      "RepositoryUsage" : {
        "type" : "object",
        "properties" : {
          "repository" : {
            "$ref" : "#/components/schemas/RepositoryRef"
          },
          "usages" : {
            "type" : "integer",
            "format" : "int64"
          },
          "modules" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ModuleUsage"
            }
          }
        }
      },
      "UsageSummary" : {
        "type" : "object",
        "properties" : {
          "usages" : {
            "type" : "integer",
            "format" : "int64"
          },
          "repositories" : {
            "type" : "integer",
            "format" : "int32"
          },
          "byRepository" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RepositoryUsage"
            }
          }
        }
      },
      "PageSymbolHit" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SymbolHit"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "SymbolHit" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "key" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "CLASS", "INTERFACE", "ENUM", "RECORD", "ANNOTATION_TYPE", "METHOD", "CONSTRUCTOR", "FIELD" ]
          },
          "display" : {
            "type" : "string"
          },
          "origin" : {
            "type" : "string",
            "enum" : [ "SOURCE", "BINARY" ]
          },
          "nameOnly" : {
            "type" : "boolean"
          },
          "usageCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "repositoryCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    }
  }
}